agent/requirements.txt
fastapi
uvicorn
langgraph
langchain-anthropic
langchain-community
langchain-core
anthropic
hvac
cryptography
pyyaml
python-dotenv
websockets
httpx
apscheduler
redis
pydantic
duckduckgo-search
requests
pytz
ddgs
